auth: fix login with new model
Web interface works again; however, client creation is still broken.
